+++ R/data.R
@@ -11,12 +11,16 @@
 #' \item{alloc_and_eff_couns}{The string name of the allocation and efficiency countries target in the drake cache, giving the countries from which final-to-useful allocations and efficiencies can be drawn for final-to-useful analyses. This is a superset of countries.}
 #' \item{max_year}{The integer representing the maximum year to be analyzed.}
 #' \item{iea_data_path}{A string representing the path to the file containing IEA extended energy balance data.}
+#' \item{ceda_data_folder}{A string representing the path to the file containing CEDA data.}
+#' \item{machine_data_path}{A string representing the path to the folder containing machine data excel files.}
 #' \item{exemplar_table_path}{A string representing the path to the exemplar table.}
 #' \item{fu_analysis_folder}{A string representing the folder containing final-to-useful analyses for various countries.}
 #' \item{report_source_folders}{A string vector representing folders containing reports to be run as the last step of the workflow.}
 #' \item{report_dest_folder}{A string representing the folder containing reports from the workflow.}
 #' \item{AllIEAData}{The name of a data frame containing all IEA data read from `iea_data_path`.}
 #' \item{IEAData}{A filtered version of `AllIEAData` containing information for only `countries`.}
+#' \item{CEDAData}{The name of the data frame containing all CEDA temperature data read from `ceda_data_folder`.}
+#' \item{AllMachineData}{A data frame containing Eta.fu and Phi.u values read through functions in `machine_functions.R`.}
 #' \item{balanced_before}{A boolean indicating whether the `IEAData` are balanced before any further analysis. They usually are not, so this value is typically `FALSE`.}
 #' \item{BalancedIEAData}{A balanced version of `IEAData`.}
 #' \item{balanced_after}{Same as `balanced_before`, only for after balancing. This should be `TRUE`.}
@@ -27,9 +31,12 @@
 #' \item{ExemplarLists}{A data frame of lists of exemplar countries for each country in `countries`, and maybe more.}
 #' \item{CompletedAllocationTables}{A data frame of completed final-to-useful allocation tables.}
 #' \item{CompletedEfficiencyTables}{A data frame of completed final-to-useful efficiency tables.}
-#' \item{AllocationGraphs}{A data frame containing final-to-useful  allocation graphs.}
+#' \item{AllocationGraphs}{A data frame containing final-to-useful allocation graphs.}
+#' \item{NonStationaryAllocationGraphs}{A data frame containing final-to-useful allocation graphs, for non-stationary data only.}
 #' \item{EfficiencyGraphs}{A data frame containing final-to-useful efficiency graphs.}
+#' \item{ExergyEnergyGraphs}{A data frame containing exergy-to-energy ratio graphs.}
 #' \item{report_source_paths}{A string vector containing paths to report files. These strings provide a detailed view of all the reports that `drake` will generate and should be all report files in `report_source_folders`.}
+#' \item{report_dest_paths}{A string for the path to a folder into which reports will written.}
 #' \item{reports_complete}{A boolean indicating success (`TRUE`) or failure (`FALSE`) of report generation.}
 #' }
 #'